Communication-Backend/requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
Pillow==8.3.2
pytz==2021.1
sqlparse==0.4.2
